Fixed the unwanted ability to add 0 donuts in DonutsActivity.java, and the unwanted ability to add an empty Order to the StoreOrders.
However, when finalizing an Order, StoreOrdersActivity will only display the list of Orders minus the most recent Order.
So, when an Order is placed in OrdersActivity, it will not show up in StoreOrdersActivity. When a second Order is placed, the first Order from before will show in StoreOrdersActivity, but not the second, and so forth. This needs to be fixed.